Stopped zeroing of floats in IEEE754tools.cpp
I experienced incorrect float-to-double conversions such as the float value 5632.20 being converted to 0x38 00 00 00 00 00 00 00 (it is supposed to be 0x40 B6 00 32 40 00 00 00). I noticed that the union _FLOATCONV f1 had all its bits reset to zero after fl.p = IEEEfloat(). Assigning the float 'number' to f1 after assigning the result of IEEEfloat() to f1 fixed the problem for me.

If it helps, I am using Arduino IDE 1.8.15 on Linux Mint 20.1 to program an Arduino Mega.